+++ app/src/main/AndroidManifest.xml
@@ -21,31 +21,31 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         tools:ignore="AllowBackup">
-        <activity android:name=".activity.ApprovalPendingFollowListActivity" />
-        <activity android:name=".activity.ApplicatedFollowListActivity" />
-        <activity android:name=".activity.AuthActivity">
+        <activity android:name=".view.ui.ApprovalPendingFollowListActivity" />
+        <activity android:name=".view.ui.ApplicatedFollowListActivity" />
+        <activity android:name=".view.ui.AuthActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".activity.CommentActivity" />
-        <activity android:name=".activity.DisplayCommentActivity" />
-        <activity android:name=".activity.EmailCheckActivity" />
-        <activity android:name=".activity.FollowerListActivity" />
-        <activity android:name=".activity.FollowingListActivity" />
+        <activity android:name=".view.ui.CommentActivity" />
+        <activity android:name=".view.ui.DisplayCommentActivity" />
+        <activity android:name=".view.ui.EmailCheckActivity" />
+        <activity android:name=".view.ui.FollowerListActivity" />
+        <activity android:name=".view.ui.FollowingListActivity" />
         <activity
-            android:name=".activity.MapActivity"
+            android:name=".view.ui.MapActivity"
             android:label="@string/title_activity_map" />
-        <activity android:name=".activity.PostActivity" />
-        <activity android:name=".activity.PostSearchActivity" />
-        <activity android:name=".activity.SettingActivity" />
-        <activity android:name=".activity.TimelineActivity" />
+        <activity android:name=".view.ui.PostActivity" />
+        <activity android:name=".view.ui.PostSearchActivity" />
+        <activity android:name=".view.ui.SettingActivity" />
+        <activity android:name=".view.ui.TimelineActivity" />
         <activity
-            android:name=".activity.UserActivity"
+            android:name=".view.ui.UserActivity"
             android:label="UserActivity" />
-        <activity android:name=".activity.UserUpdateActivity" />
+        <activity android:name=".view.ui.UserUpdateActivity" />
         <!--
             Google MapsベースのAPIのAPIキーは文字列リソースとして定義されています。
             ファイル「res/values/google_maps_api.xml」を参照）。
