+++ app/src/main/java/com/example/snap_develop/view/adapter/DisplayCommentAdapter.java
@@ -12,9 +12,9 @@ import androidx.recyclerview.widget.RecyclerView;
 import com.example.snap_develop.MainApplication;
 import com.example.snap_develop.MyDebugTree;
 import com.example.snap_develop.R;
-import com.example.snap_develop.activity.UserActivity;
 import com.example.snap_develop.bean.PostBean;
 import com.example.snap_develop.bean.UserBean;
+import com.example.snap_develop.view.ui.UserActivity;
 import com.example.snap_develop.view.viewHolder.DisplayCommentViewHolder;
 
 import java.util.List;
