+++ app/src/main/java/com/example/snap_develop/view/adapter/FollowListAdapter.java
@@ -11,8 +11,8 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.snap_develop.MyDebugTree;
 import com.example.snap_develop.R;
-import com.example.snap_develop.activity.UserActivity;
 import com.example.snap_develop.bean.UserBean;
+import com.example.snap_develop.view.ui.UserActivity;
 import com.example.snap_develop.view.viewHolder.FollowListViewHolder;
 
 import java.util.List;
