+++ app/src/main/java/com/example/snap_develop/view/adapter/PostSearchAdapter.java
@@ -13,10 +13,10 @@ import androidx.recyclerview.widget.RecyclerView;
 import com.example.snap_develop.MainApplication;
 import com.example.snap_develop.MyDebugTree;
 import com.example.snap_develop.R;
-import com.example.snap_develop.activity.DisplayCommentActivity;
-import com.example.snap_develop.activity.UserActivity;
 import com.example.snap_develop.bean.PostBean;
 import com.example.snap_develop.bean.UserBean;
+import com.example.snap_develop.view.ui.DisplayCommentActivity;
+import com.example.snap_develop.view.ui.UserActivity;
 import com.example.snap_develop.view.viewHolder.PostSearchViewHolder;
 
 import java.util.List;
