+++ app/src/main/java/com/example/snap_develop/view/ui/ApprovalPendingFollowListActivity.java
@@ -1,4 +1,4 @@
-package com.example.snap_develop.activity;
+package com.example.snap_develop.view.ui;
 
 import android.content.Intent;
 import android.graphics.Canvas;
@@ -26,8 +26,8 @@ import com.example.snap_develop.MainApplication;
 import com.example.snap_develop.MyDebugTree;
 import com.example.snap_develop.R;
 import com.example.snap_develop.bean.UserBean;
-import com.example.snap_develop.databinding.ActivityFollowingListBinding;
-import com.example.snap_develop.view.adapter.FollowListAdapter;
+import com.example.snap_develop.databinding.ActivityApprovalPendingFollowListBinding;
+import com.example.snap_develop.view.adapter.ApprovalPendingFollowListAdapter;
 import com.example.snap_develop.viewModel.FollowViewModel;
 import com.example.snap_develop.viewModel.UserViewModel;
 import com.google.android.material.tabs.TabLayout;
@@ -36,12 +36,12 @@ import java.util.List;
 
 import timber.log.Timber;
 
-public class FollowingListActivity extends AppCompatActivity implements TabLayout.OnTabSelectedListener {
+public class ApprovalPendingFollowListActivity extends AppCompatActivity implements TabLayout.OnTabSelectedListener {
 
     private UserViewModel mUserViewModel;
     private FollowViewModel mFollowViewModel;
-    private ActivityFollowingListBinding mBinding;
-    private FollowListAdapter mFollowListAdapter;
+    private ActivityApprovalPendingFollowListBinding mBinding;
+    private ApprovalPendingFollowListAdapter mApprovalPendingFollowListAdapter;
     private RecyclerView mRecyclerView;
     private List<UserBean> mFollowList;
     private String mUid;
@@ -50,40 +50,40 @@ public class FollowingListActivity extends AppCompatActivity implements TabLayou
     protected void onCreate(Bundle savedInstanceState) {
         Timber.i(MyDebugTree.START_LOG);
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_following_list);
-        setTitle("フォロー");
+        setContentView(R.layout.activity_approval_pending_follow_list);
+        setTitle("フォロー承認待ち");
 
         mFollowViewModel = new ViewModelProvider(this).get(FollowViewModel.class);
         mUserViewModel = new ViewModelProvider(this).get(UserViewModel.class);
-        mBinding = DataBindingUtil.setContentView(this, R.layout.activity_following_list);
+        mBinding = DataBindingUtil.setContentView(this, R.layout.activity_approval_pending_follow_list);
+
+        mBinding.listTabLayout.getTabAt(0).select();
 
         mBinding.buttonTabLayout.getTabAt(MainApplication.USER_POS).select();
         mBinding.buttonTabLayout.getTabAt(MainApplication.USER_POS).getIcon().setColorFilter(
                 ContextCompat.getColor(this, R.color.colorPrimary), PorterDuff.Mode.SRC_IN);
 
+        mBinding.listTabLayout.addOnTabSelectedListener(this);
         mBinding.buttonTabLayout.addOnTabSelectedListener(this);
 
-        // フォローリストを取得したら
+        // フォロー承認待ちリストを取得したら
         mFollowViewModel.getFollowList().observe(this, new Observer<List<UserBean>>() {
             @Override
             public void onChanged(List<UserBean> followList) {
                 Timber.i(MyDebugTree.START_LOG);
                 Timber.i(String.format("%s %s=%s", MyDebugTree.INPUT_LOG, "followList", followList));
 
-                mFollowListAdapter = new FollowListAdapter(FollowingListActivity.this, followList);
-                mRecyclerView = mBinding.followingRecyclerView;
-                LinearLayoutManager llm = new LinearLayoutManager(FollowingListActivity.this);
+                mApprovalPendingFollowListAdapter = new ApprovalPendingFollowListAdapter(
+                        ApprovalPendingFollowListActivity.this, followList);
+                mRecyclerView = mBinding.approvalPendingFollowRecyclerView;
+                LinearLayoutManager llm = new LinearLayoutManager(ApprovalPendingFollowListActivity.this);
                 mRecyclerView.setLayoutManager(llm);
-                RecyclerView.ItemDecoration itemDecoration = new DividerItemDecoration(FollowingListActivity.this,
-                        DividerItemDecoration.VERTICAL);
+                RecyclerView.ItemDecoration itemDecoration = new DividerItemDecoration(
+                        ApprovalPendingFollowListActivity.this, DividerItemDecoration.VERTICAL);
                 mRecyclerView.addItemDecoration(itemDecoration);
-                mRecyclerView.setAdapter(mFollowListAdapter);
+                mRecyclerView.setAdapter(mApprovalPendingFollowListAdapter);
 
                 mFollowList = followList;
-
-                if (!mUid.equals(mUserViewModel.getCurrentUser().getUid())) {
-                    return;
-                }
 
                 ItemTouchHelper.SimpleCallback callback = new ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.LEFT) {
                     @Override
@@ -95,9 +95,9 @@ public class FollowingListActivity extends AppCompatActivity implements TabLayou
                     @Override
                     public void onSwiped(@NonNull RecyclerView.ViewHolder viewHolder, int direction) {
                         int swipedPosition = viewHolder.getAdapterPosition();
-                        cancelFollowing(mFollowList.get(swipedPosition).getUid(), mUid);
+                        cancelFollow(mFollowList.get(swipedPosition).getUid(), mUid);
                         listRemove(swipedPosition);
-                        Toast.makeText(getApplication(), "フォローを解除しました", Toast.LENGTH_SHORT);
+                        Toast.makeText(getApplication(), "フォロー申請をキャンセルしました", Toast.LENGTH_SHORT);
                     }
 
                     @Override
@@ -137,15 +137,8 @@ public class FollowingListActivity extends AppCompatActivity implements TabLayou
             }
         });
 
-        // 他人のユーザー情報を表示する時（uidがIntentに設定されている時）
-        mUid = getIntent().getStringExtra("uid");
-        if (mUid == null || mUid.equals(mUserViewModel.getCurrentUser().getUid())) {
-            // 自分のユーザー情報を表示する時（uidがIntentに設定されていない時）
-            mUid = mUserViewModel.getCurrentUser().getUid();
-        }
-        Timber.i(String.format("%s=%s", "mUid", mUid));
-
-        mFollowViewModel.fetchFollowingList(mUid);
+        mUid = mUserViewModel.getCurrentUser().getUid();
+        mFollowViewModel.fetchApprovalPendingList(mUid);
     }
 
 
@@ -156,32 +149,46 @@ public class FollowingListActivity extends AppCompatActivity implements TabLayou
     }
 
 
-    private void cancelFollowing(String uid, String myUid) {
-        mFollowViewModel.deleteFollower(/*from*/uid, /*delete*/myUid);
-        mFollowViewModel.deleteFollowing(myUid, uid);
+    private void cancelFollow(String uid, String myUid) {
+        mFollowViewModel.deleteApplicatedFollow(/*from*/uid, /*delete*/myUid);
+        mFollowViewModel.deleteApprovalPendingFollow(myUid, uid);
     }
 
 
     private void listRemove(int position) {
         mFollowList.remove(position);
-        mBinding.followingRecyclerView.getAdapter().notifyItemRemoved(position);
-        mBinding.followingRecyclerView.getAdapter().notifyItemRangeRemoved(position, mFollowList.size());
+        mBinding.approvalPendingFollowRecyclerView.getAdapter().notifyItemRemoved(position);
+        mBinding.approvalPendingFollowRecyclerView.getAdapter().notifyItemRangeRemoved(position, mFollowList.size());
     }
 
 
     @Override
     public void onTabSelected(TabLayout.Tab tab) {
         Timber.i(MyDebugTree.START_LOG);
-        switch (tab.getPosition()) {
-            case MainApplication.TIMELINE_POS:
-                startActivity(new Intent(getApplication(), TimelineActivity.class));
-                break;
-            case MainApplication.MAP_POS:
-                startActivity(new Intent(getApplication(), MapActivity.class));
-                break;
-            case MainApplication.USER_POS:
-                startActivity(new Intent(getApplication(), UserActivity.class));
-                break;
+        int i = tab.parent.getId();
+        Timber.i("parent = " + tab.parent + " pos = " + tab.getPosition());
+
+        if (i == R.id.listTabLayout) {
+            switch (tab.getPosition()) {
+                case 0:
+                    startActivity(new Intent(getApplication(), ApprovalPendingFollowListActivity.class));
+                    break;
+                case 1:
+                    startActivity(new Intent(getApplication(), ApplicatedFollowListActivity.class));
+                    break;
+            }
+        } else if (i == R.id.buttonTabLayout) {
+            switch (tab.getPosition()) {
+                case MainApplication.TIMELINE_POS:
+                    startActivity(new Intent(getApplication(), TimelineActivity.class));
+                    break;
+                case MainApplication.MAP_POS:
+                    startActivity(new Intent(getApplication(), MapActivity.class));
+                    break;
+                case MainApplication.USER_POS:
+                    startActivity(new Intent(getApplication(), UserActivity.class));
+                    break;
+            }
         }
     }
 
@@ -193,16 +200,31 @@ public class FollowingListActivity extends AppCompatActivity implements TabLayou
     @Override
     public void onTabReselected(TabLayout.Tab tab) {
         Timber.i(MyDebugTree.START_LOG);
-        switch (tab.getPosition()) {
-            case MainApplication.TIMELINE_POS:
-                startActivity(new Intent(getApplication(), TimelineActivity.class));
-                break;
-            case MainApplication.MAP_POS:
-                startActivity(new Intent(getApplication(), MapActivity.class));
-                break;
-            case MainApplication.USER_POS:
-                startActivity(new Intent(getApplication(), UserActivity.class));
-                break;
+        int i = tab.parent.getId();
+
+        Timber.i("parent = " + tab.parent + " pos = " + tab.getPosition());
+
+        if (i == R.id.listTabLayout) {
+            switch (tab.getPosition()) {
+                case 0:
+                    startActivity(new Intent(getApplication(), ApprovalPendingFollowListActivity.class));
+                    break;
+                case 1:
+                    startActivity(new Intent(getApplication(), ApplicatedFollowListActivity.class));
+                    break;
+            }
+        } else if (i == R.id.buttonTabLayout) {
+            switch (tab.getPosition()) {
+                case MainApplication.TIMELINE_POS:
+                    startActivity(new Intent(getApplication(), TimelineActivity.class));
+                    break;
+                case MainApplication.MAP_POS:
+                    startActivity(new Intent(getApplication(), MapActivity.class));
+                    break;
+                case MainApplication.USER_POS:
+                    startActivity(new Intent(getApplication(), UserActivity.class));
+                    break;
+            }
         }
     }
 }
