+++ app/src/main/java/com/example/snap_develop/view/ui/MapActivity.java
@@ -1,4 +1,4 @@
-package com.example.snap_develop.activity;
+package com.example.snap_develop.view.ui;
 
 import android.Manifest;
 import android.annotation.SuppressLint;
