+++ app/src/main/java/com/example/snap_develop/view/ui/UserUpdateActivity.java
@@ -1,4 +1,4 @@
-package com.example.snap_develop.activity;
+package com.example.snap_develop.view.ui;
 
 import android.app.Activity;
 import android.content.Intent;
@@ -9,7 +9,9 @@ import android.os.Bundle;
 import android.provider.MediaStore;
 import android.text.TextUtils;
 import android.view.View;
+import android.widget.Toast;
 
+import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.content.ContextCompat;
 import androidx.databinding.DataBindingUtil;
@@ -19,31 +21,22 @@ import androidx.lifecycle.ViewModelProvider;
 import com.example.snap_develop.MainApplication;
 import com.example.snap_develop.MyDebugTree;
 import com.example.snap_develop.R;
-import com.example.snap_develop.bean.PostBean;
-import com.example.snap_develop.databinding.ActivityPostBinding;
-import com.example.snap_develop.viewModel.MapViewModel;
-import com.example.snap_develop.viewModel.PostViewModel;
+import com.example.snap_develop.bean.UserBean;
+import com.example.snap_develop.databinding.ActivityUserupdateBinding;
 import com.example.snap_develop.viewModel.UserViewModel;
-import com.google.android.gms.location.FusedLocationProviderClient;
-import com.google.android.gms.location.LocationServices;
-import com.google.android.gms.maps.model.LatLng;
 import com.google.android.material.tabs.TabLayout;
 
 import java.io.IOException;
-import java.util.Date;
-
-import javax.annotation.Nullable;
 
 import timber.log.Timber;
 
-public class PostActivity extends AppCompatActivity implements View.OnClickListener, TabLayout.OnTabSelectedListener {
+public class UserUpdateActivity extends AppCompatActivity implements View.OnClickListener,
+        TabLayout.OnTabSelectedListener {
 
     private UserViewModel mUserViewModel;
-    private PostViewModel mPostViewModel;
-    private MapViewModel mMapViewModel;
-    private ActivityPostBinding mBinding;
-    private FusedLocationProviderClient mFusedLocationClient;
-    private String mPhotoName = "";
+    private ActivityUserupdateBinding mBinding;
+    private String mUid;
+    private String mIconName;
     private Bitmap mBitMap;
     private static final int REQUEST_GALLERY = 0;
 
@@ -51,71 +44,85 @@ public class PostActivity extends AppCompatActivity implements View.OnClickListe
     protected void onCreate(Bundle savedInstanceState) {
         Timber.i(MyDebugTree.START_LOG);
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_post);
-        setTitle("投稿");
+        setContentView(R.layout.activity_userupdate);
+        setTitle("ユーザー情報更新");
 
-        mPostViewModel = new ViewModelProvider(this).get(PostViewModel.class);
-        mMapViewModel = new ViewModelProvider(this).get(MapViewModel.class);
         mUserViewModel = new ViewModelProvider(this).get(UserViewModel.class);
-        mBinding = DataBindingUtil.setContentView(this, R.layout.activity_post);
+        mBinding = DataBindingUtil.setContentView(this, R.layout.activity_userupdate);
 
-        mBinding.buttonTabLayout.getTabAt(MainApplication.MAP_POS).select();
-        mBinding.buttonTabLayout.getTabAt(MainApplication.MAP_POS).getIcon().setColorFilter(
+        mBinding.buttonTabLayout.getTabAt(MainApplication.USER_POS).select();
+        mBinding.buttonTabLayout.getTabAt(MainApplication.USER_POS).getIcon().setColorFilter(
                 ContextCompat.getColor(this, R.color.colorPrimary), PorterDuff.Mode.SRC_IN);
 
-        mBinding.postButton.setOnClickListener(this);
-        mBinding.photoImageButton.setOnClickListener(this);
+        mBinding.updateIconImageButton.setOnClickListener(this);
+        mBinding.updateButton.setOnClickListener(this);
         mBinding.buttonTabLayout.addOnTabSelectedListener(this);
 
-        // 端末の位置を取得したら投稿処理をし地図画面に遷移する
-        mMapViewModel.getDeviceLatLng().observe(this, new Observer<LatLng>() {
+        //ユーザー情報の変化が完了した後の処理
+        mUserViewModel.getUpdateResult().observe(this, new Observer<String>() {
             @Override
-            public void onChanged(@Nullable final LatLng latLng) {
+            public void onChanged(@Nullable final String updateResult) {
                 Timber.i(MyDebugTree.START_LOG);
-                Timber.i(String.format("%s %s=%s", MyDebugTree.INPUT_LOG, "latLng", latLng));
-
-                PostBean postBean = new PostBean();
-                postBean.setMessage(String.valueOf(mBinding.postTextInputEditText.getText()));
-                postBean.setPhotoName(mPhotoName);
-                postBean.setPhoto(mBitMap);
-                postBean.setLatLng(latLng);
-                postBean.setDatetime(new Date());
-                postBean.setAnonymous(mBinding.anonymousSwitch.isChecked());
-                postBean.setDanger(mBinding.dangerSwitch.isChecked());
-                postBean.setUid(mUserViewModel.getCurrentUser().getUid());
-                postBean.setType("post");
-
-                mPostViewModel.insertPost(postBean);
-                startActivity(new Intent(getApplication(), MapActivity.class));
+                Timber.i(String.format("%s %s=%s", MyDebugTree.INPUT_LOG, "updateResult", updateResult));
+                Toast.makeText(UserUpdateActivity.this, "ユーザー情報を更新しました", Toast.LENGTH_LONG).show();
+                startActivity(new Intent(getApplication(), UserActivity.class));
             }
         });
+
+        mUid = mUserViewModel.getCurrentUser().getUid();
+        mUserViewModel.fetchUserInfo(mUid);
+        mBinding.setUserViewModel(mUserViewModel);
+        mBinding.setLifecycleOwner(this);
     }
 
 
-    private void insertPost() {
+    private void updateUser() {
         Timber.i(MyDebugTree.START_LOG);
-        String post = mBinding.postTextInputEditText.getText().toString();
-        if (!validateForm(post)) {
+        //ユーザー情報を更新する処理
+        String name = mBinding.updateNameTextInputEditText.getText().toString();
+        String profile = mBinding.updateProfileTextInputEditText.getText().toString();
+
+        if (!validateForm(name, profile)) {
             return;
         }
 
-        mFusedLocationClient = LocationServices.getFusedLocationProviderClient(this);
-        mMapViewModel.fetchDeviceLocation(mFusedLocationClient);
+        UserBean updateBean = new UserBean();
+        updateBean.setUid(mUid);
+        updateBean.setName(name);
+        updateBean.setMessage(profile);
+        if (mBitMap == null) {
+            mBitMap = mUserViewModel.getUser().getValue().getIcon();
+            mIconName = mUserViewModel.getUser().getValue().getIconName();
+        }
+        updateBean.setIcon(mBitMap);
+        updateBean.setIconName(mIconName);
+
+        mUserViewModel.updateUser(updateBean);
     }
 
 
-    private boolean validateForm(String post) {
+    private boolean validateForm(String name, String profile) {
         Timber.i(MyDebugTree.START_LOG);
-        Timber.i(String.format("%s %s=%s", MyDebugTree.INPUT_LOG, "post", post));
+        Timber.i(String.format("%s %s=%s, %s=%s", MyDebugTree.INPUT_LOG, "name", name, "profile", profile));
         boolean isValidSuccess = false;
 
-        int MAX_LENGTH = 200;
-        if (TextUtils.isEmpty(post)) {
-            mBinding.postTextInputLayout.setError("コメントを入力してください");
-        } else if (post.length() > MAX_LENGTH) {
-            mBinding.postTextInputLayout.setError("コメントは200文字以内にしてください");
+        final int MAX_NAME_LENGTH = 30;
+        if (TextUtils.isEmpty(name)) {
+            mBinding.updateNameTextInputEditText.setError("ユーザー名を入力してください");
+        } else if (name.length() > MAX_NAME_LENGTH) {
+            mBinding.updateNameTextInputEditText.setError("ユーザー名は30文字以内にしてください");
         } else {
-            mBinding.postTextInputLayout.setError(null);
+            mBinding.updateNameTextInputEditText.setError(null);
+            isValidSuccess = true;
+        }
+
+        final int MAX_PROFILE_LENGTH = 200;
+        if (TextUtils.isEmpty(profile)) {
+            mBinding.updateProfileTextInputEditText.setError("プロフィールを入力してください");
+        } else if (name.length() > MAX_PROFILE_LENGTH) {
+            mBinding.updateNameTextInputEditText.setError("プロフィールは200文字以内にしてください");
+        } else {
+            mBinding.updateProfileTextInputEditText.setError(null);
             isValidSuccess = true;
         }
 
@@ -134,10 +141,11 @@ public class PostActivity extends AppCompatActivity implements View.OnClickListe
 
 
     @Override
-    protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
+    protected void onActivityResult(int requestCode, int resultCode, @javax.annotation.Nullable Intent data) {
         Timber.i(MyDebugTree.START_LOG);
         Timber.i(
-                String.format("%s %s=%s, %s=%s, %s=%s", MyDebugTree.INPUT_LOG, "requestCode", requestCode, "resultCode",
+                String.format("%s %s=%s, %s=%s, %s=%s", MyDebugTree.INPUT_LOG, "requestCode", requestCode,
+                        "resultCode",
                         resultCode, "Intent", data));
 
         super.onActivityResult(requestCode, resultCode, data);
@@ -148,10 +156,10 @@ public class PostActivity extends AppCompatActivity implements View.OnClickListe
             Timber.i("画像取得");
             if (data != null) {
                 Uri uri = data.getData();
-                mPhotoName = uri.getLastPathSegment();
+                mIconName = uri.getLastPathSegment();
                 try {
                     mBitMap = MediaStore.Images.Media.getBitmap(getContentResolver(), uri);
-                    mBinding.photoImageButton.setImageBitmap(mBitMap);
+                    mBinding.updateIconImageButton.setImageBitmap(mBitMap);
                 } catch (IOException e) {
                     e.printStackTrace();
                 }
@@ -165,10 +173,10 @@ public class PostActivity extends AppCompatActivity implements View.OnClickListe
         Timber.i(MyDebugTree.START_LOG);
         int i = view.getId();
         Timber.i(getResources().getResourceEntryName(i));
-        if (i == R.id.postButton) {
-            insertPost();
-        } else if (i == R.id.photoImageButton) {
+        if (i == R.id.updateIconImageButton) {
             pickPhoto();
+        } else if (i == R.id.updateButton) {
+            updateUser();
         }
     }
