+++ app/src/main/res/layout/activity_applicated_follow_list.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".activity.ApplicatedFollowListActivity">
+        tools:context=".view.ui.ApplicatedFollowListActivity">
 
         <com.google.android.material.tabs.TabLayout
             android:id="@+id/listTabLayout"
