+++ app/src/main/res/layout/activity_approval_pending_follow_list.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".activity.ApprovalPendingFollowListActivity">
+        tools:context=".view.ui.ApprovalPendingFollowListActivity">
 
         <com.google.android.material.tabs.TabLayout
             android:id="@+id/buttonTabLayout"
