+++ app/src/main/res/layout/activity_auth.xml
@@ -6,7 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".activity.AuthActivity">
+        tools:context=".view.ui.AuthActivity">
 
         <TextView
             android:id="@+id/snapTextView"
