+++ app/src/main/res/layout/activity_comment.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".activity.CommentActivity">
+        tools:context=".view.ui.CommentActivity">
 
         <Switch
             android:id="@+id/commentAnonymousSwitch"
