+++ app/src/main/res/layout/activity_email_check.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".activity.EmailCheckActivity">
+        tools:context=".view.ui.EmailCheckActivity">
 
         <TextView
             android:id="@+id/emailCheckTextView"
