+++ app/src/main/res/layout/activity_map.xml
@@ -7,7 +7,7 @@
         android:id="@+id/ConstraintLayout"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".activity.MapActivity">
+        tools:context=".view.ui.MapActivity">
 
         <View
             android:id="@+id/divider"
