+++ app/src/main/res/layout/activity_post.xml
@@ -6,7 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".activity.PostActivity">
+        tools:context=".view.ui.PostActivity">
 
         <ImageButton
             android:id="@+id/photoImageButton"
