+++ app/src/main/res/layout/activity_post_search.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".activity.PostSearchActivity">
+        tools:context=".view.ui.PostSearchActivity">
 
         <com.google.android.material.tabs.TabLayout
             android:id="@+id/buttonTabLayout"
