+++ app/src/main/res/layout/activity_setting.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".activity.SettingActivity">
+        tools:context=".view.ui.SettingActivity">
 
         <RadioGroup
             android:id="@+id/publicationRadioGroup"
